Added save function to file-io, with tests
Instead of dumping to a MemoryStream, and then piping to a file, we
dump to two separate files, and compare their hashes. We need to
modify the file to clear out the "start_time" field because db.dump
adds a timestamp which would invalidate the hash check. We use
readFile to do this instead of streams, because it was easier.